+++ src/main/java/org/javlo/component/container/IContainer.java
@@ -32,5 +32,7 @@ public interface IContainer extends IContentVisualComponent {
 	public IContainer getCloseComponent(ContentContext ctx);
 	
 	public IContainer getOpenComponent(ContentContext ctx);
+	
+	public String getXHTMLCode(ContentContext ctx);
 
 }